test1 (#12)
* Delete .github/workflows/github-actions-demo.yml

* unfinished tests.yml

* Add files via upload (#3)

* Update and rename tests.yml to publish-to-pypi.yml

* Create python-publish.yml (#4)

* Delete .github/workflows/publish-to-pypi.yml (#5)

* Add files via upload

* Delete malware_examples-0.0.1.tar.gz

* Add files via upload

* Delete code/__init__.py (#6)

* Delete code/harry.py (#7)

* Delete code/requirements.txt (#8)

* Delete code/voldemort.py (#9)

* Create .txt (#10)

* Add files via upload

* Delete code/.txt

* Create /code/ransomware

* Add files via upload

* Delete code/ransomware/.txt

* Delete malware_examples-0.0.2.tar.gz (#11)

* Add files via upload
