+++ code/__init__.py
@@ -1,2 +1,5 @@
-from .voldemort import ransomware_encrypt
-from .harry import ransomware_decrypt
+import sys
+from .ransomware import ransomware_encrypt
+from .ransomware import ransomware_decrypt
+
+sys.path.append("/home/abhic00l/my_project/malware_examples/ransomware")
