+++ code/requirements.txt
@@ -1,0 +1,71 @@
+alabaster==0.7.13
+Babel==2.13.1
+bcrypt==4.0.1
+bleach==5.0.1
+cached-property==1.5.1
+certifi==2023.7.22
+cffi==1.16.0
+chardet==5.1.0
+charset-normalizer==3.3.1
+cmarkgfm==0.8.0
+cryptography==41.0.5
+DateTime==5.2
+docutils==0.20.1
+email==6.0.0a1
+html5lib==1.1
+idna==3.4
+imagesize==1.4.1
+importlib-metadata==6.8.0
+jaraco.classes==3.3.0
+jeepney==0.8.0
+Jinja2==3.1.2
+keyboard==0.13.5
+keyring==24.2.0
+lockfile==0.12.2
+malware-examples==0.0.2
+markdown-it-py==3.0.0
+MarkupSafe==2.1.3
+mdit-py-plugins==0.4.0
+mdurl==0.1.2
+more-itertools==10.1.0
+myst-parser==2.0.0
+nh3==0.2.14
+olefile==0.46
+packaging==23.2
+paramiko==2.4.2
+Pillow==9.4.0
+pkginfo==1.9.6
+pyasn1==0.5.0
+pycparser==2.21
+Pygments==2.16.1
+PyNaCl==1.5.0
+PySide2==5.13.0
+python-daemon==2.2.3
+python-xlib==0.33
+pytz==2023.3.post1
+pyxhook==1.0.0
+PyYAML==6.0.1
+readme-renderer==42.0
+requests==2.31.0
+requests-toolbelt==1.0.0
+rfc3986==2.0.0
+rich==13.6.0
+roman==3.3
+scp==0.13.2
+SecretStorage==3.3.3
+shiboken2==5.13.0
+six==1.16.0
+snowballstemmer==2.2.0
+Sphinx==7.2.6
+sphinxcontrib-applehelp==1.0.7
+sphinxcontrib-devhelp==1.0.5
+sphinxcontrib-htmlhelp==2.0.4
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.6
+sphinxcontrib-serializinghtml==1.1.9
+tinycss2==1.2.1
+twine==4.0.2
+urllib3==2.0.7
+webencodings==0.5.1
+zipp==3.17.0
+zope.interface==6.1
