+++ setup.py
@@ -1,16 +1,16 @@
 from setuptools import setup
 
 __project__ = "malware_examples"
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 __description__ = "a few basic examples of malware along with decryptors"
-__packages__ = ["malware_example"]
+__packages__ = ["malware_examples"]
 __author__ = "Abhimanyu Daripally"
 __author_email__ = "abhi.cl9898@gmail.com"
 __classifiers__ = [
     "Development Status :: 2 - Pre-Alpha",
     "Intended Audience :: Education",
     "Programming Language :: Python :: 3.11",
-    "License :: Free for non-commercial use",
+    "License :: OSI Approved :: MIT License",
     "Natural Language :: English",
     "Operating System :: Other OS",
